requirements.txt
Flask==3.0.3
gunicorn==22.0.0
line-bot-sdk==3.9.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.30
